refactor(ssh): hoist the OpenSSH transport into @agentbox/sandbox-core
hetzner and digitalocean carried near-verbatim copies of ssh-cli.ts and
ssh-tunnel.ts, already drifting (DO had added keepalives and a namespaced
per-box ssh dir). remote-docker would have been a third copy, so move the
canonical implementation into sandbox-core and re-export from both.

Generalized on the way, for the remote-docker target shape:
- identity / knownHosts / user are now optional. A provider that rides the
  user's own ~/.ssh/config must inherit its identity, port and username;
  emitting -i or UserKnownHostsFile would override exactly that.
- port is emitted as `-o Port=` so ssh and scp (which spells it -P) agree.
- SshTunnelManager takes a namespace; defaultBoxSshDir sanitizes the box id
  (remote-docker's sandbox-id embeds an SSH destination, so it has @ : and /).

Also register the remote-docker provider in the config layer: the PROVIDERS
row, the per-provider image/size/checkpoint keys, and box.remoteDockerHost.
providerKeyCap now camelizes across a hyphen — box.imageRemote-docker is not
a legal key.
