+++ packages/config/schema/user-config.schema.json
@@ -18,6 +18,7 @@
         "defaultCheckpointVercel": { "type": "string", "minLength": 1 },
         "defaultCheckpointE2b": { "type": "string", "minLength": 1 },
         "defaultCheckpointDigitalocean": { "type": "string", "minLength": 1 },
+        "defaultCheckpointRemoteDocker": { "type": "string", "minLength": 1 },
         "size": { "type": "string", "minLength": 1 },
         "sizeDocker": { "type": "string", "minLength": 1 },
         "sizeDaytona": { "type": "string", "minLength": 1 },
@@ -25,6 +26,7 @@
         "sizeVercel": { "type": "string", "minLength": 1 },
         "sizeE2b": { "type": "string", "minLength": 1 },
         "sizeDigitalocean": { "type": "string", "minLength": 1 },
+        "sizeRemoteDocker": { "type": "string", "minLength": 1 },
         "withPlaywright": { "type": "boolean" },
         "claudeInstall": { "enum": ["native", "npm"] },
         "withEnv": { "type": "boolean" },
@@ -42,6 +44,7 @@
         "imageVercel": { "type": "string", "minLength": 1 },
         "imageE2b": { "type": "string", "minLength": 1 },
         "imageDigitalocean": { "type": "string", "minLength": 1 },
+        "imageRemoteDocker": { "type": "string", "minLength": 1 },
         "dockerCacheShared": { "type": "boolean" },
         "memory": { "type": "integer", "minimum": 0 },
         "cpus": { "type": "integer", "minimum": 0 },
@@ -51,6 +54,7 @@
         "hetznerLocation": { "type": "string", "minLength": 1 },
         "digitaloceanRegion": { "type": "string", "minLength": 1 },
         "digitaloceanProject": { "type": "string", "minLength": 1 },
+        "remoteDockerHost": { "type": "string", "minLength": 1 },
         "inbound": { "type": "string", "minLength": 1 },
         "vercelTimeoutMs": { "type": "integer", "minimum": 1 },
         "vercelNetworkPolicy": { "type": "string" },
